Renamed persistence servcice to favoriteComics
